API/requirements.txt
pandas==2.0.3
numpy==1.24.3
# sentence-transformers 2.2.2 expects cached_download from huggingface_hub
sentence-transformers==2.2.2
transformers==4.24.0
huggingface-hub==0.13.3
faiss-cpu==1.7.4
scikit-learn==1.3.0
# PyTorch 2.0.1 is fine here
torch==2.0.1
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
# Avoid urllib3 v2 OpenSSL warnings on macOS LibreSSL
urllib3<2

